refactored unittests to use the utility class functions and pass the parameters instead of building the jsonparameters Removed dead unittets moved unittests from testaimsunbridge to importnetwork
