+++ test/TMG.Aimsun.Tests/TestModuleImportNetwork.cs
@@ -19,7 +19,6 @@
 
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Newtonsoft.Json;
-using System.IO;
 
 namespace TMG.Aimsun.Tests
 {
@@ -29,27 +28,32 @@ namespace TMG.Aimsun.Tests
         [TestMethod]
         public void ImportNetwork()
         {
+            string networkPath = Helper.BuildFilePath("inputFiles\\Frabitztown.nwp");
             string modulePath = Helper.BuildModulePath("inputOutput\\importNetwork.py");
-            string jsonParameters = JsonConvert.SerializeObject(new
-            {
-                NetworkPackageFile = Helper.BuildFilePath("inputFiles\\Frabitztown.nwp")
-            });
-            Helper.Modeller.Run(null, modulePath, jsonParameters);
+            Utility.RunImportNetworkTool(networkPath, modulePath);
         }
 
         [TestMethod]
         public void TestSaveNetwork()
         {
+            string networkPath = Helper.BuildFilePath("inputFiles\\Frabitztown.nwp");
             string modulePath = Helper.BuildModulePath("inputOutput\\importNetwork.py");
-            string jsonParameters = JsonConvert.SerializeObject(new
-            {
-                NetworkPackageFile = Helper.BuildFilePath("inputFiles\\Frabitztown.nwp")
-
-            });
-            Helper.Modeller.Run(null, modulePath, jsonParameters);
+            Utility.RunImportNetworkTool(networkPath, modulePath);
 
             //build an output file location of where to save the file
             string outputPath = Helper.BuildFilePath("aimsunFiles\\test3\\FrabitztownNetwork.ang");
+            Helper.Modeller.SaveNetworkModel(null, outputPath);
+        }
+
+        [TestMethod]
+        public void TestSaveNetworkWithInvalidPath()
+        {
+            string networkPath = Helper.BuildFilePath("inputFiles\\Frabitztown.nwp");
+            string modulePath = Helper.BuildModulePath("inputOutput\\importNetwork.py");
+            Utility.RunImportNetworkTool(networkPath, modulePath);
+
+            //build an output file location of where to save the file
+            string outputPath = Helper.BuildFilePath("aimsunFiles\\test3\\BadPath.ang");
             Helper.Modeller.SaveNetworkModel(null, outputPath);
         }
     }
